My first personal class work
This project was given to all #codeplateau 2.0 team 2021 to work on as a class work by the DG (Daser David) on 24th may, 2021. It is basically for us to create page containing the following; about us, home, contact us, products, and services. This project will gradually be worked on as time permits.